+++ libraries/transformer/src/main/java/androidx/media3/transformer/Transformer.java
@@ -1475,6 +1475,7 @@ public final class Transformer {
     if (!doesFormatsMatch(mediaItemInfo, firstEditedMediaItem)) {
       remuxingMuxerWrapper = null;
       transformerInternal = null;
+      exportResultBuilder.reset();
       exportResultBuilder.setOptimizationResult(OPTIMIZATION_FAILED_FORMAT_MISMATCH);
       processFullInput();
       return;